Iteration 6
The AI now responds to player proximity.

The NPC will either move towards the target, and stop once close, or
flee, and resume normal bahaviour once out of range.

Error in code, the NPC will slerp to the exact opposite of the desired
look rotation. Unsure how to fix, but will be useful for flee
bahaviour!!