Iteration 10.0.6
DRE script seems to have the closest resemblance to Dead Reckoning. It
works nicely if it can guess where the player is headed(i.e. in a
straight line) but it plays up a little when you add more complex
movements in sequence like running around in random directions.

Still needs work but is closer to the final goal.

P.S: DRY2 and Dead Reckoning scripts are basically useless and are
seperate attempts with different code.