Iteration 11: Room selection
1. Basic implementation for different rooms.

Game starts on menu like scene, with ability to create multiple rooms
based on name. Will need to elaborate and add more functionality like
options for volume etc.

Created gameobject & scripts to handle transfer of room name, which are
subsequently deleted when no longer of use.

Also, music added in first scene!!

2. Small cleanup in Scenes folder. Only 2 scenes are currently in use,
the others are redundant, and thus placed in an appropriate folder.

3. Dead Reckoning is still in it's early stages, will need to speak to
Hamid or Warren about why the hell it ain't cooperating.